+++ tamcmc/headers/models.h
@@ -37,7 +37,8 @@ VectorXd model_MS_Global_a1etaa3_AppWidth_HarveyLike_v2(const VectorXd& params, 
 VectorXd model_MS_local_basic(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false);
 VectorXd model_MS_local_Hnlm(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false); // Added on 14 Feb 2020 : Handles local fit without fitting inclination... instead used Hnlm
 
-VectorXd model_Test_Gaussian(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false);
+//VectorXd model_Test_Gaussian(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false);
+VectorXd model_Kallinger2014_Gaussian(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false);
 VectorXd model_Harvey_Gaussian(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false);
 
 //VectorXd model_RGB_asympt_a1etaa3_AppWidth_HarveyLike(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false); 
@@ -55,7 +56,10 @@ VectorXd model_MS_Global_a1a2a3_HarveyLike(const VectorXd& params, const VectorX
 VectorXd model_MS_Global_aj_HarveyLike(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams=false); // Added on 18 Jan 2021: Handles the a2 coefficient with 2 terms : a constant term + 1 slope term in nu
 VectorXd model_MS_Global_ajAlm_HarveyLike(const VectorXd& params, const VectorXi& params_length, const VectorXd& x, bool outparams, external_data); // Added on 31 Mar 2021
 
-void write_star_params(const VectorXd& spec_params, const VectorXd& raw_params, const VectorXi& plength, const MatrixXd& mode_params, const MatrixXd& noise_params, const std::string file_out, const std::string modelname, const std::string name_params);
+void write_star_params(const VectorXd& spec_params, const VectorXd& raw_params, const VectorXi& plength, 
+            const MatrixXd& mode_params, const MatrixXd& noise_params, const std::string file_out, 
+            const std::string modelname, const std::string name_params, 
+            const std::string noise_name_params="# Input mode parameters. H0 , tau_0 , p0 / H1, tau_1, p1 / N0. Set at -1 if not used. -2 means that the parameter is not even written on the file (because irrelevant).");
 void write_star_params_mixed(const VectorXd& spec_params, const VectorXd& raw_params, const VectorXi& plength, const MatrixXd& mode_params, const MatrixXd& noise_params, 
     const std::string file_out, const std::string modelname, const std::string name_params, const MatrixXd& mixed_mode_params, const std::string mixed_mode_name_params,
     const VectorXd global_mixed_mode_params, const std::vector<std::string> global_mixed_mode_name_params, VectorXd nu_p, VectorXd nu_g);