+++ tamcmc/sources/noise_models.cpp
@@ -6,13 +6,11 @@
  */
 #include <math.h>
 #include <Eigen/Dense>
-//#include <iostream>
-//#include <iomanip>
+#include <iostream>
+#include <fstream>
+#include <iomanip>
 
 using Eigen::VectorXd;
-
-//VectorXd harvey_like(const VectorXd noise_params, const VectorXd& x, const VectorXd& y, const int Nharvey);
-//VectorXd harvey1985(const VectorXd noise_params, const VectorXd& x, const VectorXd& y, const int Nharvey);
 
 VectorXd harvey_like(const VectorXd& noise_params, const VectorXd& x, const VectorXd& y, const int Nharvey){
 	/* This function calculate a sum of harvey like profile + a white noise and adds 
@@ -67,5 +65,91 @@ VectorXd harvey1985(const VectorXd& noise_params, const VectorXd& x, const Vecto
 	y_out=y_out + white_noise;
 
 return y_out;
+}
+
+double get_ksinorm(const double b, const double c, const Eigen::VectorXd& x) {
+    double integral = 0.0;
+    double h = x(1) - x(0); // assuming x is equally spaced
+    
+    for (int i = 0; i < x.size(); i++) {
+        double term = 1.0 / (1.0 + std::pow(x(i) / b, c));
+        
+        if (i == 0 || i == x.size() - 1) {
+            integral += 0.5 * term;
+        } else {
+            integral += term;
+        }
+    }
+    
+    integral *= h;
+    double ksi = b/integral;
+    
+    return ksi;
+}
+
+
+VectorXd eta_squared_Kallinger2014(const VectorXd& x){
+	const double x_nyquist=x.maxCoeff();
+	VectorXd eta(x.size());
+	eta=sin(0.5*M_PI*x.array()/x_nyquist)/(0.5*M_PI*x.array()/x_nyquist);
+	if (x[0] == 0){ // This to avoid the Division by 0
+		eta[0]=1;
+	}
+	return eta.array().square();
+}
+
+VectorXd Kallinger2014(const double numax, const double mu_numax, const double Mass, const VectorXd& noise_params,const VectorXd& x, const VectorXd& y){
+/*
+	Using notations from Table 2 of Kallinger+2014 (https://arxiv.org/pdf/1408.0817.pdf)
+	Not that here we assume the instrumental noise to be Pinstrument(nu) = 0
+	The noise_params must have parameters in this order:
+		- Noise a : ka, sa, t
+		- Noise b1: k1, s1, ( and c1, the slope of the SuperLorentzian)
+		- Noise b2: k2, s2, ( and c2, the slope of the SuperLorentzian)
+	Such that at the end we have: [ka,sa,t,k1,s1,c1, k2,s2,c2, N0]
+*/
+	const long Nx=x.size();
+	VectorXd ones(Nx), white_noise(Nx), tmp0(Nx), tmp1(Nx), tmp2(Nx), y0(Nx), Power(Nx);
+	ones.setOnes();
+	// Compute the Leakage effect as a sinc function (Eq 1 of Kallinger+2014)
+	const VectorXd eta_squared=eta_squared_Kallinger2014(x);
+	// Compute b1, b2 and a
+	const double a=std::abs(noise_params[0])*std::pow(std::abs(numax + mu_numax),noise_params[1]) * std::pow(Mass,noise_params[2]);
+	const double b1=std::abs(noise_params[3]*std::pow(std::abs(numax + mu_numax),noise_params[4]));
+	const double b2=std::abs(noise_params[6]*std::pow(std::abs(numax + mu_numax),noise_params[7]));
+	const double c1=std::abs(noise_params[5]);
+	const double c2=std::abs(noise_params[8]);
+	// Compute the normalisation constants ksi1 and ksi2
+	const double ksi1=get_ksinorm(b1, c1, x);
+	const double ksi2=get_ksinorm(b2, c2, x);
+	y0=y;
+	// White noise first, added to the input y-vector
+	white_noise.setConstant(noise_params.tail(1)(0));
+ 	Power=y + white_noise;
+	// First SuperLorentzian
+	tmp0=(x/b1).array().pow(c1); // Denominator
+	tmp1=(eta_squared * ksi1 * std::pow(a,2)/b1).cwiseProduct((tmp0 + ones).cwiseInverse()); // Numerator/Denominator
+	Power=Power + tmp1;
+	// Second SuperLorentzian
+	tmp0=(x/b2).array().pow(c2); // Denominator
+	tmp2= (eta_squared *ksi2*std::pow(a,2)/b2).cwiseProduct((tmp0 + ones).cwiseInverse()); // Numerator/Denominator
+	Power=Power + tmp2;
+	/*
+    std::ofstream debugFile("debug.txt"); // Open the debug file for writing
+	debugFile << "!a=" << a << std::endl;
+	debugFile << "!b1=" << b1 << std::endl;
+	debugFile << "!b2=" << b2 << std::endl;
+	debugFile << "!c1=" << c1 << std::endl;
+	debugFile << "!c2=" << c2 << std::endl;
+	debugFile << "!ksi1=" << ksi1 << std::endl;
+	debugFile << "!ksi2=" << ksi2 << std::endl;
+    debugFile << "#x       model   eta_squared   P1   P2    N0   y_in" << std::endl;
+    for (int i = 0; i < x.size(); i++) {
+        debugFile << x[i] << "\t " << Power[i]  << "\t " << eta_squared[i]  << "\t " << tmp1[i] << "\t " << tmp2[i]  << "\t " << white_noise[i] << "\t " << y0[i] << std::endl;
+    }
+    debugFile.close();
+	*/
+	return Power;
+
 }